issue #360: get username from cookies
