+++ package.json
@@ -33,6 +33,7 @@
     "i18next-http-backend": "^2.6.2",
     "i18next-resources-to-backend": "^1.2.1",
     "jest-environment-jsdom": "^29.7.0",
+    "js-cookie": "^3.0.5",
     "next": "14.2.15",
     "react": "^18",
     "react-dom": "^18",
@@ -51,6 +52,7 @@
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.14",
+    "@types/js-cookie": "^3.0.6",
     "@types/node": "^20",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.1",
