+++ public/locales/en/homePage.json
@@ -1,23 +1,23 @@
 {
-    "submit_button": "Submit",
-    "submit_button_disabled_hint": "You need to add a minimum of 1 file to start the analysis",
-    "fileList":{
-        "uploadedfiles": "Uploaded Files",
-        "noUploadedfiles": "No uploaded files"
-    },
-    "fileElement": {
-        "altText": {
-            "uploadedFileAlt": "Uploaded file picture",
-            "deleteFileAlt": "Delete file button icon"
-        }
-    },
-    "dropzone": {
-        "altText": {
-            "hoveredImageAlt": "Hovered image",
-            "CloudIconAlt": "File upload icon"
-        },
-        "dragDrop": "Drag and Drop Files",
-        "or": "OR",
-        "browseFile": "Browse Files"
+  "submitButton": "Submit",
+  "submitButtonDisabledHint": "You need to add a minimum of 1 file to start the analysis",
+  "fileList": {
+    "uploadedfiles": "Uploaded Files",
+    "noUploadedfiles": "No uploaded files"
+  },
+  "fileElement": {
+    "altText": {
+      "uploadedFileAlt": "Uploaded file picture",
+      "deleteFileAlt": "Delete file button icon"
     }
+  },
+  "dropzone": {
+    "altText": {
+      "hoveredImageAlt": "Hovered image",
+      "CloudIconAlt": "File upload icon"
+    },
+    "dragDrop": "Drag and Drop Files",
+    "or": "OR",
+    "browseFile": "Browse Files"
+  }
 }
