+++ src/app/api/extract-label-data/route.ts
@@ -1,5 +1,5 @@
+import { handleApiError } from "@/utils/server/apiErrors";
 import { pipelineApi } from "@/utils/server/backend";
-import { handleApiError } from "@/utils/server/common";
 
 export async function POST(request: Request) {
   const formData = await request.formData();
