+++ src/app/api/signup/route.ts
@@ -1,7 +1,7 @@
+import { handleApiError } from "@/utils/server/apiErrors";
 import { usersApi } from "@/utils/server/backend";
 
 export async function POST(request: Request) {
-
   const authHeader = request.headers.get("Authorization");
   if (!authHeader) {
     return new Response(
@@ -12,18 +12,12 @@ export async function POST(request: Request) {
     );
   }
 
-  return usersApi.signupSignupPost({headers: { Authorization: authHeader }}).then((response) => {
-    return Response.json(response.data);
-  }).catch((error) => {
-    if (error.response) {
-      console.error("Error response:", error.response.data);
-    } else if (error.request) {
-      console.error("Error request:", error.request);
-    } else {
-      console.error("Error message:", error.message);
-    }
-    return new Response(JSON.stringify({ error: error.message }), {
-      status: error.status,
+  return usersApi
+    .signupSignupPost({ headers: { Authorization: authHeader } })
+    .then((response) => {
+      return Response.json(response.data);
+    })
+    .catch((error) => {
+      return handleApiError(error);
     });
-  });
 }
