+++ src/app/label-data-validation/[id]/page.tsx
@@ -3,9 +3,10 @@ import LabelDataValidator from "@/components/LabelDataValidator";
 import useAlertStore from "@/stores/alertStore";
 import useUploadedFilesStore from "@/stores/fileStore";
 import { DEFAULT_LABEL_DATA } from "@/types/types";
-import { mapInspectionToLabelData } from "@/utils/client/common";
+import { mapInspectionToLabelData } from "@/utils/client/modelTransformation";
 import { Inspection } from "@/utils/server/backend";
 import axios from "axios";
+import Cookies from "js-cookie";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 import { validate } from "uuid";
@@ -27,7 +28,7 @@ export default function Page({ params }: { params: { id: string } }) {
       return;
     }
 
-    const username = "";
+    const username = atob(Cookies.get("token") ?? "");
     const password = "";
     const authHeader = "Basic " + btoa(`${username}:${password}`);
     const controller = new AbortController();
